Fixed warnings, commented out some parts.

+++ Frameworks/HFHexFiend/HFProgressTracker.h
@@ -37,7 +37,7 @@
     unsigned long long maxProgress;
     UIProgressView *progressIndicator;
     NSTimer *progressTimer;
-    double lastSetValue;
+    float lastSetValue;
     NSDictionary *userInfo;
     id delegate;
 }
